Creation of ICARUSTriggerV2 fragment to store new trigger information and restore ability of decoder to store all data. ICARUSTriggerUDP has been reverted to how it was in release v1_00_01 to work with old data before trigger system upgrades
