+++ sbndaq-artdaq-core/Overlays/ICARUS/ICARUSTriggerV2Fragment.cc
@@ -1,0 +1,44 @@
+#include "sbndaq-artdaq-core/Overlays/ICARUS/ICARUSTriggerV2Fragment.hh"
+#include "cetlib_except/exception.h"
+
+
+
+std::ostream & icarus::operator << (std::ostream & os, ICARUSTriggerV2Fragment const & f) {
+ 
+  os << "ICARUSTriggerV2Fragment: " << f.GetDataString();
+  os << "\n\t Event Number: " << f.getWREventNo();
+  os << "\n\t Seconds since UTC epoch: " << f.getWRSeconds();
+  os << "\n\t Nanoseconds: " << f.getWRNanoSeconds();
+  os << "\n\t Gate ID: " << f.getGateID();
+  os << "\n\t Gate Type: " << f.getGateType();
+  os << "\n\t isBNB: " << f.isBNB();
+  os << "\n\t isNuMI: " << f.isNuMI();
+  os << "\n\t Last Overall Timestamp: " << f.getLastTimestamp();
+  os << "\n\t Last BNB Timestamp: " << f.getLastTimestampBNB();
+  os << "\n\t Last NuMI Timestamp: " << f.getLastTimestampNuMI();
+  os << "\n\t Delta Gates: " << f.getDeltaGates();
+  os << "\n\t Delta Gates BNB: " << f.getDeltaGatesBNB();
+  os << "\n\t Delta Gates NuMI: " << f.getDeltaGatesNuMI();
+  
+  os << std::endl;
+  return os;
+}
+/*
+std::ostream & icarus::operator << (std::ostream & os, ICARUSTriggerV2FragmentMetadata const & m) {
+  os << "\n\t Name: " << m.getName();
+  os << "\n\t Event Number: " << m.getEventNo();
+  os << "\n\t Seconds: " << m.getSeconds();
+  os << "\n\t Nanoseconds: " << m.getNanoSeconds();
+  os << "\n\t WR: " << m.getWRName();
+  os << "\n\t WR Event Number: " << m.getWREventNo();
+  os << "\n\t WR Seconds: " << m.getWRSeconds();
+  os << "\n\t WR Nanoseconds: " << m.getWRNanoSeconds();
+  os << std::endl;
+  return os;
+}
+*/
+
+
+bool icarus::ICARUSTriggerV2Fragment::Verify() const {
+  return true;
+}
